refactor: 抽离 HttpSender 出站 HTTP 发送（E2 第一步）
OrzEasyBot 的出站 HTTP 发送 + 健康状态机（sendToTarget/sendBatch/批量结果解析/幂等键/
背压 Semaphore/并发计数聚合）抽到 HttpSender，OrzEasyBot 只保留路由 + WS + 入站解析。
834 → ~530 行。test + integrationTest 全绿。
